requirements.txt
-i https://pypi.org/simple
bcrypt==4.3.0; python_version >= '3.8'
cffi==1.17.1; platform_python_implementation != 'PyPy'
cmd2==2.7.0; python_version >= '3.9'
cryptography==45.0.4; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
future==1.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
netmiko==4.5.0; python_version >= '3.9' and python_version < '4.0'
ntc-templates==7.9.0; python_version >= '3.8' and python_version < '4.0'
paramiko==3.5.1; python_version >= '3.6'
pycparser==2.22; python_version >= '3.8'
pygments==2.19.2; python_version >= '3.8'
pynacl==1.5.0; python_version >= '3.6'
pyperclip==1.9.0
pyserial==3.5
python-gnupg==0.5.4
pyyaml==6.0.2; python_version >= '3.8'
rich==14.1.0; python_full_version >= '3.8.0'
rich-argparse==1.7.1; python_version >= '3.8'
ruamel.yaml==0.18.14; python_version >= '3.8'
ruamel.yaml.clib==0.2.12; python_version < '3.14' and platform_python_implementation == 'CPython'
scp==0.15.0
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
textfsm==1.1.3
wcwidth==0.2.13
